docs(scripts): describe remaining alias reconciliation
