+++ docs/quality/file-size-baseline.json
@@ -1,6 +1,6 @@
 {
   "_comment": [
-    "Baseline of Rust source files that exceed the SPECIFICATION §14.1 500-line",
+    "Baseline of Rust source files that exceed the SPECIFICATION \u00a714.1 500-line",
     "budget. Recorded so the mandatory rule can be enforced going forward without",
     "a mass file split: files listed here are grandfathered at their recorded line",
     "count and must not grow by 10% or more; files not listed here must stay at or",
@@ -13,13 +13,9 @@
   "limit": 500,
   "growthAllowance": 0.1,
   "files": {
-    "rust/src/builtins/builtin_word_definitions.rs": 1398,
-    "rust/src/coreword_registry.rs": 1334,
-    "rust/src/types/value_operations.rs": 1321,
     "rust/src/arithmetic_operation_tests.rs": 834,
     "rust/src/types/fraction_mcdc_tests.rs": 789,
     "rust/src/wasm_interpreter_bindings/wasm_value_conversion.rs": 741,
-    "rust/src/types/mod.rs": 722,
     "rust/src/interpreter/interpreter_core.rs": 682,
     "rust/src/types/fraction.rs": 681,
     "rust/src/interpreter/arithmetic.rs": 656,
