+++ rust/src/builtins/generated_core_word_docs.rs
@@ -2,6 +2,7 @@
 
 pub(crate) struct GeneratedCoreWordDoc {
     pub(crate) name: &'static str,
+    pub(crate) summary: &'static str,
     pub(crate) hover_summary: &'static str,
     pub(crate) hover_syntax: &'static str,
 }
@@ -9,346 +10,415 @@ pub(crate) struct GeneratedCoreWordDoc {
 pub(crate) const GENERATED_CORE_WORD_DOCS: &[GeneratedCoreWordDoc] = &[
     GeneratedCoreWordDoc {
         name: "EAT",
+        summary: "Set the consumption mode to consume operands.",
         hover_summary: "EAT — consume operands",
         hover_syntax: ", +",
     },
     GeneratedCoreWordDoc {
         name: "KEEP",
+        summary: "Set the consumption mode to keep operands.",
         hover_summary: "KEEP — keep operands and append result",
         hover_syntax: ",, +",
     },
     GeneratedCoreWordDoc {
         name: "GET",
+        summary: "Extract one element of a vector by index.",
         hover_summary: "GET — extract element at index",
         hover_syntax: "[ 10 20 30 ] [ 0 ] GET",
     },
     GeneratedCoreWordDoc {
         name: "INSERT",
+        summary: "Insert a value at a given index in a vector.",
         hover_summary: "INSERT — insert element at index",
         hover_syntax: "[ 1 3 ] [ 1 2 ] INSERT",
     },
     GeneratedCoreWordDoc {
         name: "REPLACE",
+        summary: "Replace an element of a vector at a given index.",
         hover_summary: "REPLACE — replace element at index",
         hover_syntax: "[ 1 2 3 ] [ 0 9 ] REPLACE",
     },
     GeneratedCoreWordDoc {
         name: "REMOVE",
+        summary: "Remove an element from a vector at a given index.",
         hover_summary: "REMOVE — remove element at index",
         hover_syntax: "[ 1 2 3 ] [ 0 ] REMOVE",
     },
     GeneratedCoreWordDoc {
         name: "LENGTH",
+        summary: "Return the number of elements in a vector.",
         hover_summary: "LENGTH — return element count",
         hover_syntax: "[ 1 2 3 ] LENGTH",
     },
     GeneratedCoreWordDoc {
         name: "TAKE",
+        summary: "Take the first N or last -N elements of a vector.",
         hover_summary: "TAKE — take N elements from start or end",
         hover_syntax: "[ 1 2 3 4 5 ] [ 3 ] TAKE",
     },
     GeneratedCoreWordDoc {
         name: "SPLIT",
+        summary: "Split a vector into chunks at the specified sizes.",
         hover_summary: "SPLIT — split vector at sizes",
         hover_syntax: "[ 1 2 3 4 ] [ 2 2 ] SPLIT",
     },
     GeneratedCoreWordDoc {
         name: "CONCAT",
+        summary: "Flatten and concatenate two vectors.",
         hover_summary: "CONCAT — flatten and concatenate vectors",
         hover_syntax: "[ 1 2 ] [ 3 4 ] CONCAT",
     },
     GeneratedCoreWordDoc {
         name: "REVERSE",
+        summary: "Reverse the order of vector elements.",
         hover_summary: "REVERSE — reverse element order",
         hover_syntax: "[ 1 2 3 ] REVERSE",
     },
     GeneratedCoreWordDoc {
         name: "RANGE",
+        summary: "Generate a numeric sequence from a [start, end] pair.",
         hover_summary: "RANGE — generate numeric sequence",
         hover_syntax: "[ 0 5 ] RANGE",
     },
     GeneratedCoreWordDoc {
         name: "REORDER",
+        summary: "Reorder vector elements according to an index permutation.",
         hover_summary: "REORDER — reorder by index list",
         hover_syntax: "[ 'a' 'b' 'c' ] [ 2 0 1 ] REORDER",
     },
     GeneratedCoreWordDoc {
         name: "COLLECT",
+        summary: "Collect N items off the stack into a new vector.",
         hover_summary: "COLLECT — collect N items into vector",
         hover_syntax: "1 2 3 3 COLLECT",
     },
     GeneratedCoreWordDoc {
         name: "TRUE",
+        summary: "Push the boolean TRUE onto the stack.",
         hover_summary: "TRUE — push TRUE",
         hover_syntax: "TRUE",
     },
     GeneratedCoreWordDoc {
         name: "FALSE",
+        summary: "Push the boolean FALSE onto the stack.",
         hover_summary: "FALSE — push FALSE",
         hover_syntax: "FALSE",
     },
     GeneratedCoreWordDoc {
         name: "NIL",
+        summary: "Push the NIL value onto the stack.",
         hover_summary: "NIL — push NIL",
         hover_syntax: "NIL",
     },
     GeneratedCoreWordDoc {
         name: "NIL?",
+        summary: "Test whether the top value is an operational NIL (absent).",
         hover_summary: "NIL? — test whether a value is absent",
         hover_syntax: "1 0 / NIL?",
     },
     GeneratedCoreWordDoc {
         name: "NIL-REASON",
+        summary: "Read the direct reason of an operational NIL as a protocol-string Text.",
         hover_summary: "NIL-REASON — read the NIL reason protocol string",
         hover_syntax: "1 0 / NIL-REASON",
     },
     GeneratedCoreWordDoc {
         name: "CHARS",
+        summary: "Split a string into a vector of one-character strings.",
         hover_summary: "CHARS — split string into characters",
         hover_syntax: "'hi' CHARS",
     },
     GeneratedCoreWordDoc {
         name: "JOIN",
+        summary: "Join a vector of strings into a single string.",
         hover_summary: "JOIN — join characters into string",
         hover_syntax: "[ 'h' 'i' ] JOIN",
     },
     GeneratedCoreWordDoc {
         name: "TRIM",
+        summary: "Remove whitespace from both ends of a string.",
         hover_summary: "TRIM — strip leading and trailing whitespace",
         hover_syntax: "'  hi  ' TRIM",
     },
     GeneratedCoreWordDoc {
         name: "TOKENIZE",
+        summary: "Split a string into a vector of substrings using a separator.",
         hover_summary: "TOKENIZE — split string by separator",
         hover_syntax: "'a,b,c' ',' TOKENIZE",
     },
     GeneratedCoreWordDoc {
         name: "SUBSTITUTE",
+        summary: "Replace every occurrence of a substring with another.",
         hover_summary: "SUBSTITUTE — replace substring occurrences",
         hover_syntax: "'hello' 'l' 'L' SUBSTITUTE",
     },
     GeneratedCoreWordDoc {
         name: "STARTS-WITH?",
+        summary: "Test whether a string begins with the given prefix.",
         hover_summary: "STARTS-WITH? — prefix predicate",
         hover_syntax: "'hello' 'he' STARTS-WITH?",
     },
     GeneratedCoreWordDoc {
         name: "ENDS-WITH?",
+        summary: "Test whether a string ends with the given suffix.",
         hover_summary: "ENDS-WITH? — suffix predicate",
         hover_syntax: "'hello' 'lo' ENDS-WITH?",
     },
     GeneratedCoreWordDoc {
         name: "NUM",
+        summary: "Parse text as a number; Bubble/NIL on parse failure.",
         hover_summary: "NUM — parse to number",
         hover_syntax: "'42' NUM",
     },
     GeneratedCoreWordDoc {
         name: "STR",
+        summary: "Convert a value to its string representation.",
         hover_summary: "STR — convert to string",
         hover_syntax: "42 STR",
     },
     GeneratedCoreWordDoc {
         name: "CHR",
+        summary: "Convert a numeric character code to a single-character string.",
         hover_summary: "CHR — make a character",
         hover_syntax: "65 CHR",
     },
     GeneratedCoreWordDoc {
         name: "ADD",
+        summary: "Add two numeric values, element-wise with broadcasting.",
         hover_summary: "ADD — add values",
         hover_syntax: "1 2 +",
     },
     GeneratedCoreWordDoc {
         name: "SUB",
+        summary: "Subtract two numeric values, element-wise with broadcasting.",
         hover_summary: "SUB — subtract values",
         hover_syntax: "5 3 -",
     },
     GeneratedCoreWordDoc {
         name: "MUL",
+        summary: "Multiply two numeric values, element-wise with broadcasting.",
         hover_summary: "MUL — multiply values",
         hover_syntax: "2 4 *",
     },
     GeneratedCoreWordDoc {
         name: "DIV",
+        summary: "Divide two numeric values exactly (fractional result).",
         hover_summary: "DIV — divide values",
         hover_syntax: "10 2 /",
     },
     GeneratedCoreWordDoc {
         name: "EQ",
+        summary: "Test equality of two values.",
         hover_summary: "EQ — test equality",
         hover_syntax: "1 1 =",
     },
     GeneratedCoreWordDoc {
         name: "LT",
+        summary: "Test less-than comparison.",
         hover_summary: "LT — test less than",
         hover_syntax: "1 2 <",
     },
     GeneratedCoreWordDoc {
         name: "LTE",
+        summary: "Test less-than-or-equal comparison.",
         hover_summary: "LTE — test less than or equal",
         hover_syntax: "1 1 <=",
     },
     GeneratedCoreWordDoc {
         name: "GT",
+        summary: "Test greater-than comparison.",
         hover_summary: "GT — test greater than",
         hover_syntax: "2 1 >",
     },
     GeneratedCoreWordDoc {
         name: "GTE",
+        summary: "Test greater-than-or-equal comparison.",
         hover_summary: "GTE — test greater than or equal",
         hover_syntax: "1 1 >=",
     },
     GeneratedCoreWordDoc {
         name: "NEQ",
+        summary: "Test inequality of two values.",
         hover_summary: "NEQ — test inequality",
         hover_syntax: "1 2 <>",
     },
     GeneratedCoreWordDoc {
         name: "AND",
+        summary: "Logical AND. A NIL operand passes through.",
         hover_summary: "AND — logical AND",
         hover_syntax: "TRUE TRUE &",
     },
     GeneratedCoreWordDoc {
         name: "OR",
+        summary: "Logical OR. A NIL operand passes through.",
         hover_summary: "OR — logical OR",
         hover_syntax: "TRUE FALSE OR",
     },
     GeneratedCoreWordDoc {
         name: "NOT",
+        summary: "Logical negation. A NIL operand passes through.",
         hover_summary: "NOT — logical negation",
         hover_syntax: "TRUE NOT",
     },
     GeneratedCoreWordDoc {
         name: "COND",
+        summary: "Evaluate guard/body clauses in order, executing the first match.",
         hover_summary: "COND — evaluate guard/body clauses",
         hover_syntax: "1 { TRUE } { 'y' } { IDLE } { 'n' } COND",
     },
     GeneratedCoreWordDoc {
         name: "VENT",
+        summary: "Lazy NIL-coalescing control directive: keep a non-NIL top and skip the following source unit; on a NIL top, discard it and evaluate the following source unit as the fallback.",
         hover_summary: "VENT — lazy NIL-coalescing fallback",
         hover_syntax: "NIL ^ [ 0 ]",
     },
     GeneratedCoreWordDoc {
         name: "MAP",
+        summary: "Apply a code block to each element of a vector.",
         hover_summary: "MAP — apply block to each element",
         hover_syntax: "[ 1 2 3 ] { [ 2 ] * } MAP",
     },
     GeneratedCoreWordDoc {
         name: "FILTER",
+        summary: "Keep only the elements for which a predicate block returns TRUE.",
         hover_summary: "FILTER — keep elements matching predicate",
         hover_syntax: "[ 1 2 3 ] { [ 2 ] = } FILTER",
     },
     GeneratedCoreWordDoc {
         name: "FOLD",
+        summary: "Reduce a vector to a single value using an initial accumulator and combiner block.",
         hover_summary: "FOLD — reduce with initial value",
         hover_syntax: "[ 1 2 3 ] [ 0 ] { + } FOLD",
     },
     GeneratedCoreWordDoc {
         name: "ANY",
+        summary: "TRUE if at least one element satisfies the predicate.",
         hover_summary: "ANY — true if any element matches",
         hover_syntax: "[ 1 2 3 ] { [ 2 ] = } ANY",
     },
     GeneratedCoreWordDoc {
         name: "ALL",
+        summary: "TRUE if every element satisfies the predicate.",
         hover_summary: "ALL — true if all elements match",
         hover_syntax: "[ 2 4 ] { [ 2 ] MOD [ 0 ] = } ALL",
     },
     GeneratedCoreWordDoc {
         name: "PRINT",
+        summary: "Write the top stack value to the output stream, consuming it. A string is written as its raw text, without the quotes the stack shows ('TEST' prints as TEST); nested strings keep their quotes.",
         hover_summary: "PRINT — output value to display",
         hover_syntax: "42 PRINT",
     },
     GeneratedCoreWordDoc {
         name: "DEF",
+        summary: "Define a user word from a body and a name.",
         hover_summary: "DEF — define user word",
         hover_syntax: "{ 2 * } 'DOUBLE' DEF",
     },
     GeneratedCoreWordDoc {
         name: "DEL",
+        summary: "Delete a user word from the dictionary.",
         hover_summary: "DEL — delete user word",
         hover_syntax: "{ [ 1 ] } 'W' DEF 'W' DEL",
     },
     GeneratedCoreWordDoc {
         name: "LOOKUP",
+        summary: "Display the documentation for a named word.",
         hover_summary: "LOOKUP — show word documentation",
         hover_syntax: "'ADD' ?",
     },
     GeneratedCoreWordDoc {
         name: "FILL",
+        summary: "Fill a target shape with a constant value.",
         hover_summary: "FILL — fill shape with value",
         hover_syntax: "[ 2 2 0 ] FILL",
     },
     GeneratedCoreWordDoc {
         name: "MOD",
+        summary: "Modulo (remainder) of two numeric values.",
         hover_summary: "MOD — modulo",
         hover_syntax: "7 3 %",
     },
     GeneratedCoreWordDoc {
         name: "FLOOR",
+        summary: "Round toward negative infinity.",
         hover_summary: "FLOOR — round toward negative infinity",
         hover_syntax: "[ 7/3 ] FLOOR",
     },
     GeneratedCoreWordDoc {
         name: "CEIL",
+        summary: "Round toward positive infinity.",
         hover_summary: "CEIL — round toward positive infinity",
         hover_syntax: "[ 7/3 ] CEIL",
     },
     GeneratedCoreWordDoc {
         name: "ROUND",
+        summary: "Round to nearest integer (half-up).",
         hover_summary: "ROUND — round to nearest integer",
         hover_syntax: "[ 5/2 ] ROUND",
     },
     GeneratedCoreWordDoc {
         name: "EXEC",
+        summary: "Execute a vector as Ajisai code.",
         hover_summary: "EXEC — execute vector as code",
         hover_syntax: "[ 1 2 + ] EXEC",
     },
     GeneratedCoreWordDoc {
         name: "SQRT",
+        summary: "Exact square root of a non-negative rational.",
         hover_summary: "SQRT — exact square root",
         hover_syntax: "2 SQRT",
     },
     GeneratedCoreWordDoc {
         name: "ABS",
+        summary: "Absolute value of a number.",
         hover_summary: "Absolute value of a number.",
         hover_syntax: "-2 ABS",
     },
     GeneratedCoreWordDoc {
         name: "NEG",
+        summary: "Numeric negation.",
         hover_summary: "Numeric negation.",
         hover_syntax: "2 NEG",
     },
     GeneratedCoreWordDoc {
         name: "SIGN",
+        summary: "Sign of a number: -1, 0, or 1.",
         hover_summary: "Sign of a number: -1, 0, or 1.",
         hover_syntax: "-2 SIGN",
     },
     GeneratedCoreWordDoc {
         name: "MIN",
+        summary: "Smaller of two numbers.",
         hover_summary: "Smaller of two numbers.",
         hover_syntax: "1 2 MIN",
     },
     GeneratedCoreWordDoc {
         name: "MAX",
+        summary: "Larger of two numbers.",
         hover_summary: "Larger of two numbers.",
         hover_syntax: "1 2 MAX",
     },
     GeneratedCoreWordDoc {
         name: "SORT",
+        summary: "Return a copy of a vector sorted in ascending order.",
         hover_summary: "Return a copy of a vector sorted in ascending order.",
         hover_syntax: "[ 3 1 2 ] SORT",
     },
     GeneratedCoreWordDoc {
         name: "UNIQUE",
+        summary: "Return a copy of a vector with duplicates removed, preserving first-occurrence order.",
         hover_summary: "UNIQUE — remove duplicates, keeping first-occurrence order",
         hover_syntax: "[ 1 2 1 ] UNIQUE",
     },
     GeneratedCoreWordDoc {
         name: "CONTAINS",
+        summary: "True if a vector contains an element equal to the given value.",
         hover_summary: "True if a vector contains an element equal to the given value.",
         hover_syntax: "[ 1 2 ] 2 CONTAINS",
     },
     GeneratedCoreWordDoc {
         name: "INDEX-OF",
+        summary: "Index of the first element equal to the value; Bubble/NIL if absent.",
         hover_summary: "Index of the first element equal to the value; Bubble/NIL if absent.",
         hover_syntax: "[ 1 2 ] 2 INDEX-OF",
     },
