+++ scripts/generate-core-word-docs.mjs
@@ -17,6 +17,7 @@ const rows = coreEntries.map((entry) => {
   if (!word) throw new Error(`No canonical contract for ${entry.canonical}`);
   return `    GeneratedCoreWordDoc {
         name: ${rustString(word.name)},
+        summary: ${rustString(word.documentation.summary)},
         hover_summary: ${rustString(word.documentation.hover)},
         hover_syntax: ${rustString(word.documentation.syntax)},
     },`;
@@ -26,6 +27,7 @@ const output = `// @generated by scripts/generate-core-word-docs.mjs; do not edi
 
 pub(crate) struct GeneratedCoreWordDoc {
     pub(crate) name: &'static str,
+    pub(crate) summary: &'static str,
     pub(crate) hover_summary: &'static str,
     pub(crate) hover_syntax: &'static str,
 }
